requirements.txt
psycopg2-binary==2.9.3
ipython==7.31.0
genanki==0.13.0
ffmpeg-python==0.2.0
